itybity
Trait FromBitIterator

pub trait FromBitIterator {
    // Required methods
    fn from_lsb0_iter(iter: impl IntoIterator<Item = bool>) -> Self;
    fn from_msb0_iter(iter: impl IntoIterator<Item = bool>) -> Self;
}

Trait used for parsing a value from a bit iterator.

Required Methods

fn from_lsb0_iter(iter: impl IntoIterator<Item = bool>) -> Self

Parses a value from an iterator of bits in Lsb0 order.

If the iterator is shorter than the number of bits in the type, the remaining bits are assumed to be zero.

fn from_msb0_iter(iter: impl IntoIterator<Item = bool>) -> Self

Parses a value from an iterator of bits in Msb0 order.

If the iterator is shorter than the number of bits in the type, the remaining bits are assumed to be zero.

Object Safety

This trait is not object safe.