itybity
Trait GetBit

pub trait GetBit<O>
where
    O: BitOrder,
{
    // Required method
    fn get_bit(&self, index: usize) -> bool;
}

Trait for getting a bit at a given index.

Required Methods

fn get_bit(&self, index: usize) -> bool

Returns the bit at the given index.

Panics

Implementations may panic if the provided index is out of bounds.