itybity
Trait IntoBitIterator

pub trait IntoBitIterator {
    type IterLsb0: Iterator<Item = bool>;
    type IterMsb0: Iterator<Item = bool>;

    // Required methods
    fn into_iter_lsb0(self) -> Self::IterLsb0;
    fn into_iter_msb0(self) -> Self::IterMsb0;

    // Provided methods
    fn into_lsb0_vec(self) -> Vec<bool>
       where Self: Sized { ... }
    fn into_msb0_vec(self) -> Vec<bool>
       where Self: Sized { ... }
}

Trait for converting iterators over values into iterators over bits.

This trait is automatically implemented for all types that implement IntoIterator where the item type implements IntoBits.

Required Associated Types

type IterLsb0: Iterator<Item = bool>

The Lsb0 bit iterator type.

type IterMsb0: Iterator<Item = bool>

The Msb0 bit iterator type.

Required Methods

fn into_iter_lsb0(self) -> Self::IterLsb0

Converts self into a bit iterator in Lsb0 order.

fn into_iter_msb0(self) -> Self::IterMsb0

Converts self into a bit iterator in Msb0 order.

Provided Methods

fn into_lsb0_vec(self) -> Vec<bool>
where Self: Sized,

Converts self into a bit vector in Lsb0 order.

fn into_msb0_vec(self) -> Vec<bool>
where Self: Sized,

Converts self into a bit vector in Msb0 order.