itybity
Trait StrToBits

pub trait StrToBits<'a> {
    // Required method
    fn iter_bits(&'a self) -> StrBitIter<'a> ;

    // Provided method
    fn to_bit_vec(&'a self) -> Vec<bool> { ... }
}

Trait for converting bit strings into a bit iterator.

Required Methods

fn iter_bits(&'a self) -> StrBitIter<'a>

Converts a bit string into a bit iterator.

The returned iterator will yield true for any character that is not '0',

Provided Methods

fn to_bit_vec(&'a self) -> Vec<bool>

Converts a bit string into a bit vector.

The returned vector will contain true for any character that is not '0',